final state before deciding to rewrite. rationale: It turns out that you cannot have a state update function for each "subject", like Participant: update holdings or Proposals: change the status. No, they must all be rolled up into one function that does everything (complex), because they change the same variable, which is 'network'.
Furthermore, the state update functions are specific to the policy that runs before it. They do not just update the state. It begs the question: should we split code by "decide what to do/do it" or "per subject" "decide what to do/do it".

Furthermore, it is not obvious what each data "looks like", for instance, I cannot know what "delta_holdings" is without reading the code that creates it. It'd be far easier if there was a definition somewhere.

Furthermore, each function cannot run independently of cadCAD. I mostly need this to look at what the data looks like, but it also hinders debugging.
Maybe I can save the state as a gzipped pickle?

Maybe you shouldn't split problems up by "subject". It might help make things simpler, but then it restricts you if you want to do more unorthodox simulation types in the future. No, you should split the problem up by entities, which are all acting upon each other. Because that's what's happening in real life.